Implement lookup_result to factorize the logic of looking at a HTML page returned by bugzilla looking for errors and then looking up a meaningfull result if no error is found. The error are handled by throwing an exception and adding an error message in the .error element. On success it returns the value extracted from the page.
The .error element is supposed to be a central place for errors that is going to be used by all states of the bug submission assistant. Although it would be better to display contextual errors, for instance next to a given field, this would lead to more complex development and most probably a user interface that require significant work to design properly.

The state_signin function is converted to use lookup_result. Because lookup_result is a filter, the http://api.jquery.com/deferred.pipe/ function is used to chain the jQuery.post and pass it to the function that will actually display the username somewhere in the page when the login is successfull. Unlike the twisted deferred, the jQuery pipe function does not catch exception and does not turn them into failures. The second argument of pipe is therefore ignored and exceptions are not caught to be passed to the next pipe.