ie specific code to disable the text part of the input[type=file] http://stackoverflow.com/questions/3914142/file-input-behaviour-in-ie explains it in more words:
I have noticed a difference in behaviour between chrome and IE of handling file input clicks.

In Chrome, clicking anywhere on the input (text or button bit) triggers the file dialogue.

In IE (testing on 7), you have to click on the actual button, not the text bit.

The problem is that I'm using a transparent file input on top of a regular text box. Clicking on the text box should open the dialog. In Chrome this is working fine, as the total size of the input is the same size as my text box.

In IE it doesn't work properly as the user needs to be clicking on the actual button part of the transparent file input.

Any ideas of how to fix this??