Shorten greeting statement for layout and readability
This commit shortens the opening "Greeting" statement in an attempt to
make it easier for people to read it quickly as their eyes move down
the screen towards the "meat" of bug reporting process.

We're also shortening the statement to make sure that it fits into the
layout properly -- on some browsers the text appears taller than the
background graphic beneath it.