Add blacklist for EOL version numbers to be withheld
This commit adds a blacklist for old, End of Life (EOL), not-active
version numbers to withhold from the BSA version picker. The picker
has about 100 entries right now, and it needs cleanup.

We don't have a hard-and-fast rule for generating this list, but
basically any version number that's way past EOL is a strong candidate
for this list. We're keeping minor version #'s around for now (3.3.0,
3.4.0), as that list isn't very large.

We aren't currently removing EOL beta builds and RC's from 3.5 and up,
however those are next on the chopping block.