EOL: Notify user when an EOL version has been selected
This commit expands on the previous EOL commit and implements a basic
EOL/unsupported version warning whenever a user selects a flagged
version in the drop-down box.

NOTE: I wasn't able to get the jquery .get() function to properly load
the data file, so the list of unsupport versions is currently
hard-coded in bug.js. When we get that issue ironed-out, we'll want to
actually read the data from that file.