Blacklist: Collapse 3.5 to "3.5 all versions"
This commit blacklists all 3.5-era versions from the BSA, replacing
them with the catch-all "3.5 all versions".

We're doing this now as we've reached the specified time (date of last
build + 6 months), and we now have a message in place indicating that
we don't support these builds but that other service providers may
continue to provide support.